requirements.txt
# --- Python Core ---
protobuf==3.20.3
numpy==1.26.4

# --- Flask & Extensions ---
Flask==2.3.3
flask-smorest==0.40.0
flask-jwt-extended==4.5.2
flask-cors==4.0.0
flask-migrate==4.0.5
Flask-SQLAlchemy==3.1.1

# --- ORM / DB ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9  # PostgreSQL 用（必要に応じて）

# --- gRPC ---
grpcio==1.44.0
grpcio-tools==1.44.0

# --- Media / ML ---
mediapipe==0.10.8
opencv-python==4.9.0.80
keras-facenet==0.3.2

# --- Utility ---
python-dotenv==1.0.1
tqdm==4.66.4
requests==2.32.3
marshmallow==3.21.1
bcrypt==4.1.3
Werkzeug==2.3.8

# --- Testing ---
pytest==8.2.1
pytest-cov==5.0.0

scipy==1.11.1
tensorflow==2.8.1